OO-716: fix rounding to int of item difficulty

+++ src/main/java/org/olat/ims/qti/statistics/manager/QTIStatisticsManagerImpl.java
@@ -302,7 +302,7 @@ public class QTIStatisticsManagerImpl implements QTIStatisticsManager {
 
 		double averageScore = totalScore / totalResults;
 		//difficulty (p-value)
-		double difficulty = numOfCorrectAnswers / totalResults;
+		double difficulty = numOfCorrectAnswers / (double)totalResults;
 		double averageDuration = totalDuration / totalResults;
 		
 		StatisticsItem stats = new StatisticsItem();
